requirements.txt
anyio==3.6.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
distlib==0.3.6
dulwich==0.21.5
exceptiongroup==1.1.1
fastapi==0.95.1
filelock==3.12.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
postgres==4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
psycopg2-pool==1.1
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.7
PyGreSQL==5.2.4
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pyscopg2==66.0.2
pytest==7.3.1
python-dotenv==1.0.0
rapidfuzz==2.15.1
requests==2.30.0
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.21.1
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0
